feat: implement auth domain layer (T072-T075)
- Add AuthRepository interface with register, login, verifyOTP, logout methods
- Implement RegisterUseCase with Vietnamese phone validation (T072)
- Implement VerifyOTPUseCase with 6-digit OTP validation (T073)
- Implement LoginUseCase with credential validation (T074)
- Implement LogoutUseCase (T075)
- All 26 unit tests passing

Tests cover:
- Valid inputs and successful operations
- Input validation (phone format, password length, OTP format)
- Error handling (invalid credentials, expired OTP)
- Whitespace trimming and phone number normalization

+++ lib/features/home/presentation/widgets/filter_bottom_sheet.dart
@@ -36,7 +36,7 @@ class _FilterBottomSheetState extends State<FilterBottomSheet> {
     final maxPrice = (filters['maxPrice'] as num?)?.toDouble() ?? _maxPrice;
     _priceRange = RangeValues(minPrice, maxPrice);
     _minRating = (filters['rating'] as num?)?.toDouble() ?? 0;
-    
+
     final conditionStr = filters['condition'] as String?;
     if (conditionStr != null) {
       _condition = ProductCondition.values.firstWhere(
@@ -91,12 +91,12 @@ class _FilterBottomSheetState extends State<FilterBottomSheet> {
                 Row(
                   children: [
                     Text(
-                      formatVND(_priceRange.start.toInt()),
+                      formatVND(_priceRange.start),
                       style: Theme.of(context).textTheme.bodyMedium,
                     ),
                     const Spacer(),
                     Text(
-                      formatVND(_priceRange.end.toInt()),
+                      formatVND(_priceRange.end),
                       style: Theme.of(context).textTheme.bodyMedium,
                     ),
                   ],
